@@ -4,7 +4,7 @@ version: 0.0.100
 homepage: https://github.com/tim-smart/elemental
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dev_dependencies:
   flutter_lints: ^2.0.0
